requirements.txt
langchain
langchain-openai
langchain_community
openai>=1.58.1,<2.0.0
pydantic==2.4.2
python-dotenv==1.0.0
fpdf2==2.7.0
requests==2.31.0
